add own submodule
